requirements.txt
asgiref==3.7.2
branca==0.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.6
django-environ==0.11.2
django-widget-tweaks==1.5.0
folium==0.15.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
numpy==1.26.2
Pillow==10.1.0
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
